Uploaded Original Code
I uploaded the original code for my senior design project. At the time, I had a very basic understanding of the python language, or coding in general. I had no experience with object-oriented programming at the time and I didn't understand proper coding techniques or styles, so the code is a bit messy to look at.